Add files via upload
Hi all,

Here is my code for the challenge. The skeleton is Mervyn's latest draft minus the random forest stuff, with a couple small things added and variable names neatened for the graphs etc. 

The preprocessing I mentioned on call which I've been working on is in there and I think it's possibly the most accurate preprocessing we can do; whether this translates into good regression scores for us is yet to be seen. 

At the moment this code is a mess - I couldn't get the "fillna" method to work properly, so naturally I settled for doing an if statement inside an if statement inside an if statement inside a for loop inside a for loop. Might need some modification.

Please use the database alongside the code to observe how we can fill values just by inspection, and any questions or advice please shoot to me.

Best,

Aakash